% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{build_gradient}
\alias{build_gradient}
\title{Build gradient}
\usage{
build_gradient(gradientlim, length)
}
\arguments{
\item{gradientlim}{double. Environmental gradient size}

\item{length}{int. Number of cells}
}
\description{
C++ code to build environmental gradients
}
\examples{
build_gradient(10, 10)

}
