% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{disperse}
\alias{disperse}
\title{Disperse}
\usage{
disperse(a, d, xmin, xmax)
}
\arguments{
\item{a}{int. Final dispesal position}

\item{d}{int. Dispersal distance}

\item{xmin}{int. Minimum cell coordinate}

\item{xmax}{int. Maximum cell coordinate}
}
\description{
C++ code for dispersal
}
\examples{
disperse(10, 10)

}
