% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{forestgapdynamics}
\alias{forestgapdynamics}
\title{Forest gap dynamics}
\usage{
forestgapdynamics(grid, Pfall, Rgaps)
}
\arguments{
\item{grid}{int.Number of cell per side of the matrix}

\item{Pfall}{double. Treefall probability}

\item{Rgaps}{int. Treefall gaps radius}
}
\description{
C++ code for forest gap dynamics
}
\examples{
forestgapdynamics(20, 2, 0.01)

}
