% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateNCI.R
\name{generateNCIsim}
\alias{generateNCIsim}
\title{Wrapper for generateNCI}
\usage{
generateNCIsim(grid = 50, Nt = 50, muNCI = 0.749, sigmaNCI = 2.651,
  p = 0.271, mu = 0.749, sigma = 2.651, lambda = 0.31, d = 3)
}
\arguments{
\item{grid}{int. Number of cells per side of the matrix.}

\item{Nt}{int. Number of time steps.}

\item{muNCI}{double. mu parameter for the normal distribution used for NCI.}

\item{sigmaNCI}{double. sigma parameter for the normal distribution used for
NCI.}

\item{p}{double. Probability to be negative used in the Bernoulli
distribution.}

\item{mu}{double. mu parameter for the lognormal distribution used for
positive deltaNCI.}

\item{sigma}{double. sigma parameter for the lognormal distribution used for
positive deltaNCI.}

\item{lambda}{double. lambda parameter for the exponential distribution used
for negative deltaNCI.}

\item{d}{int. Spatial autoccorelation size in number of cells.}
}
\value{
A matrix..
}
\description{
Format in a matrix for the simulator.
}
\examples{
generateNCI()
}
