% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateTopography.R
\name{generateTopography}
\alias{generateTopography}
\title{Generate topography}
\usage{
generateTopography(
  grid = 20,
  topography = "sinusoidal",
  Elim = 5,
  amplitude = 1,
  ruggedness = 1,
  plot = 1
)
}
\arguments{
\item{grid}{int.  Number of cells per side of the matrix.}

\item{topography}{char. Topography generator between 'sinusoidal',
'squarediamond' or 'paracou'.}

\item{Elim}{double. Environmental matrix extrme (absolute value).}

\item{amplitude}{double. Amplitude of the sinusoidal functional.}

\item{ruggedness}{double. Ruggedness parameter.}

\item{plot}{int. Plot number between 1 and 15 (a cell size is 3x3m).}
}
\value{
A matrix of TWI for the simulator.
}
\description{
Generate topography
}
\examples{
generateTopography()

}
