% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gifMaps.R
\name{gifMaps}
\alias{gifMaps}
\title{gifMaps}
\usage{
gifMaps(simulation, variable = "genotype")
}
\arguments{
\item{simulation}{df.  result from simulator function}

\item{variable}{char.  variable to visualize "topography", "gaps",
"genotype", or "phenotype"}
}
\value{
A ggplot.
}
\description{
gifMaps
}
\examples{

simulation <- simulator(grid = 50)
gifMaps(simulator())

}
