% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logit.R
\name{invlogit}
\alias{invlogit}
\title{Inverse of logistic function}
\usage{
invlogit(x)
}
\arguments{
\item{x}{double. A vector of reals.}
}
\value{
A vector of probabilities.
}
\description{
To transform reals into probabilities with exp(x) / (1 + exp(x)).
}
\examples{
invlogit(rnorm(1))

}
