% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logit.R
\name{logit}
\alias{logit}
\title{Logistic function}
\usage{
logit(p)
}
\arguments{
\item{p}{double. A vector of probability.}
}
\value{
A vector of reals.
}
\description{
To transform probabilities into reals with log( p /(1-p)).
}
\examples{
logit(pnorm(rnorm(1)))

}
