% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paracouTopography.R
\name{paracouTopography}
\alias{paracouTopography}
\title{Topography from Paracou}
\usage{
paracouTopography(plot = 1, Elim = 1)
}
\arguments{
\item{plot}{int. Plot number between 1 and 15 (a cell size is 3x3m)}

\item{Elim}{double. Environmental matrix extrme (absolute value)}
}
\value{
A matrix of TWI of ize 98x98 for the simulator.
}
\description{
Topography from Paracou
}
\examples{

paracouTopography(1)

}
