% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotEnv}
\alias{plotEnv}
\title{plotEnv}
\usage{
plotEnv(simulation, gen = 1)
}
\arguments{
\item{simulation}{df.  result from simulator function}

\item{gen}{int.  generation to be plotted}
}
\value{
A ggplot.
}
\description{
plotEnv
}
\examples{

plotEnv(simulator())

}
