% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotMaps}
\alias{plotMaps}
\title{plotMaps}
\usage{
plotMaps(simulation)
}
\arguments{
\item{simulation}{df.  result from simulator function}
}
\value{
A ggplot.
}
\description{
plotMaps
}
\examples{

plotMaps(simulator())

}
