% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotSim}
\alias{plotSim}
\title{plotSim}
\usage{
plotSim(simulation)
}
\arguments{
\item{simulation.}{df.  result from simulator function}
}
\value{
A ggplot.
}
\description{
plotSim
}
\examples{

plotSim(simulator())

}
