% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotTrajectories}
\alias{plotTrajectories}
\title{plotTrajectories}
\usage{
plotTrajectories(simulation)
}
\arguments{
\item{simulation}{df.  result from simulator function}
}
\value{
A ggplot.
}
\description{
plotTrajectories
}
\examples{

plotTrajectories(simulatorAll())

}
