% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdeltanci.R
\name{rdeltanci}
\alias{rdeltanci}
\title{The deltaNCI distribution}
\usage{
rdeltanci(
  NCI = rnorm(1, mean = 122.73, sd = 25.806),
  alpha = -1.32,
  beta = 0.003,
  mu = 0.749,
  sigma = 2.651,
  lambda = 0.31
)
}
\arguments{
\item{NCI}{double. Vector of previous NCI (size of the vector determines
number of generated deltaNCI).}

\item{alpha}{double. Intercept for the Bernoulli distribution determining the
risk to have a negative deltaNCI.}

\item{beta}{double. Slope of previous NCI for the Bernoulli distribution
determining the risk to have a negative deltaNCI.}

\item{mu}{double. mu parameter for the lognormal distribution used for
positive deltaNCI.}

\item{sigma}{double. sigma parameter for the lognormal distribution used for
positive deltaNCI.}

\item{lambda}{double. lambda parameter for the exponential distribution used
for negative deltaNCI.}
}
\value{
A vector of n generated delta nci.
}
\description{
Random generator of deltaNCI based on a proability p to be negative used in a
Bernoulli distribution. Negative deltaNCI follow an exponential distribution
of parameter lambda. Positive deltaNCI follow a lognormal distribution of
parameters mu and sigma. Default parameters correspond to values inferred in
the control and biodiversity plots of Paracou.
}
\examples{
rdeltanci(1)

}
