% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/squareDiamondTopography.R
\name{sampleM}
\alias{sampleM}
\title{Values sampling for the Square-Diamond algorithm}
\usage{
sampleM(M, x, y)
}
\arguments{
\item{M}{matrix. Environmental matrix}

\item{x}{int. Row}

\item{y}{int. Col}
}
\value{
Environmental matrix of size grid x grid
}
\description{
Special sampler of matrix returning NA if coordinates don't exists in the matrix
}
