% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sinusoidalTopography.R
\name{sinusoidalTopography}
\alias{sinusoidalTopography}
\title{Topography generator with the sinusoidal function}
\usage{
sinusoidalTopography(grid = 2^6 + 1, Elim = 10, amplitude = 1)
}
\arguments{
\item{grid}{int. Matrix grid size should be of size 2^n+1}

\item{Elim}{double. Environmental matrix extrme (absolute value)}

\item{amplitude}{double. Amplitude of the sinusoidal functional}
}
\value{
Environmental matrix of size grid x grid for the simulator
}
\description{
E = (-Elim, Elim)
M = sin(amplitude.E*t(E))
Have a look to the topography vignette.
}
\examples{
sinusoidalTopography(grid = 5, Elim = 10, amplitude = 0.05)

}
