% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/squareDiamondTopography.R
\name{square}
\alias{square}
\title{Square step from the Square-Diamond algorithm}
\usage{
square(M, x, y, hs, h = 1)
}
\arguments{
\item{M}{matrix. Environmental matrix}

\item{x}{int. Row}

\item{y}{int. Col}

\item{hs}{int. half-step}

\item{h}{double. Rudgeness}
}
\value{
Environmental matrix of size grid x grid
}
\description{
Square step from the Square-Diamond algorithm
}
