% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{disperse}
\alias{disperse}
\title{Disperse}
\usage{
disperse(a, d, xmin, xmax, ymin, ymax)
}
\arguments{
\item{a}{IntegerVector. Final dispesal positions}

\item{d}{int. Dispersal distance}

\item{xmin}{int. Minimum row}

\item{xmax}{int. Maximum row}

\item{ymin}{int. Minimum column}

\item{ymax}{int. Maximum column}
}
\description{
C++ code for dispersal
}
\examples{
disperse(c(10, 10), 2, 0, 20, 0, 20)

}
