% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{disperse}
\alias{disperse}
\title{Disperse}
\usage{
disperse(a, Rdispersal, grid)
}
\arguments{
\item{a}{IntegerVector. Final dispesal positions}

\item{Rdispersal}{int. Dispersal radius in cells}

\item{grid}{int. Number of cell per side of the matrix}
}
\description{
C++ code for dispersal
}
\examples{
disperse(c(10, 10), 2, 20)

}
