% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{forestgapdynamics}
\alias{forestgapdynamics}
\title{Forest gap dynamics}
\usage{
forestgapdynamics(grid, radius, probability)
}
\arguments{
\item{grid}{int.Number of cell per side of the matrix}

\item{radius}{int. Gaps radius}

\item{probability}{double. Gaps probability}
}
\description{
C++ code for forest gap dynamics
}
\examples{
forestgapdynamics(20, 2, 0.01)

}
