% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gifMaps.R
\name{gifMaps}
\alias{gifMaps}
\title{gifMaps}
\usage{
gifMaps(simulation, type = "genotype", variable = "topography")
}
\arguments{
\item{simulation}{df. Result from simulator function.}

\item{type}{char. Type to visualize "variable", "genotype", or
"phenotype".}

\item{variable}{char. Variable to visualize "topography" or "nci".}
}
\value{
A ggplot.
}
\description{
gifMaps
}
\examples{
gifMaps(simulatorAll(grid = 5))

}
