% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gifMaps.R
\name{gifMaps}
\alias{gifMaps}
\title{gifMaps}
\usage{
gifMaps(simulation)
}
\arguments{
\item{simulation}{df.  result from simulator function}
}
\value{
A ggplot.
}
\description{
gifMaps
}
\examples{

simulation <- simulator(grid = 50)
gifMaps(simulator())

}
