% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotSim}
\alias{plotSim}
\title{plotSim}
\usage{
plotSim(simulation, time = 1)
}
\arguments{
\item{simulation}{df.  result from simulator function}

\item{time}{int.  generation to be plotted}
}
\value{
A ggplot.
}
\description{
plotSim
}
\examples{

plotSim(simulator())

}
