% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdeltanci.R
\name{rdeltanci}
\alias{rdeltanci}
\title{The deltaNCI distribution}
\usage{
rdeltanci(n = 1, p = 0.271, mu = 0.749, sigma = 2.651,
  lambda = 0.31)
}
\arguments{
\item{n}{int. Number of observations.}

\item{p}{double. Probability to be negative used in the Bernoulli
distribution.}

\item{mu}{double. mu parameter for the lognormal distribution used for
positive deltaNCI.}

\item{sigma}{double. sigma parameter for the lognormal distribution used for
positive deltaNCI.}

\item{lambda}{double. lambda parameter for the exponential distribution used
for negative deltaNCI.}
}
\value{
A vector of n generated delta nci.
}
\description{
Random generator of deltaNCI based on a proability p to be negative used in a
Bernoulli distribution. Negative deltaNCI follow an exponential distribution
of parameter lambda. Positive deltaNCI follow a lognormal distribution of
parameters mu and sigma. Default parameters correspond to values inferred in
the control and biodiversity plots of Paracou.
}
\examples{
rdeltanci(1)

}
