% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rspcor.R
\name{rspcor}
\alias{rspcor}
\title{Spatially auto-correlated distribution}
\usage{
rspcor(
  M = matrix(rdeltanci(NCI = rnorm(50 * 50, mean = 122.73, sd = 25.806)), ncol = 50,
    nrow = 50),
  dcor = 3
)
}
\arguments{
\item{M}{matrix. Environmental matrix in which to create spatial
auto-correlation.}

\item{dcor}{int. Spatial auto-correlation size in number of cells.}
}
\value{
A data.frame of X, Y, and values.
}
\description{
Spatial random generator of spatially auto-correlated distribution based on a
given random generator. The generator uses a simple spatial resampling of a
smaller grid.
}
\examples{
rspcor()

}
