% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulator.R
\name{simulator}
\alias{simulator}
\title{simulator}
\usage{
simulator(grid = 20, Nt = 50, Elim = 5, muG = 0, sigmaG = 1,
  muE = 0, sigmaE = 1, Pfall = 0.01, Rgaps = 2, Pdeath = 0.1,
  Ns = 4, Rdispersal = 1, determinist = TRUE)
}
\arguments{
\item{grid}{int.  Number of cells per side of the matrix}

\item{Nt}{int. Number of time steps}

\item{Elim}{double. Environmental gradient size}

\item{muG}{double. Mean of genetic values}

\item{sigmaG}{double. Variance of genetic values}

\item{muE}{double. Mean of environmental values}

\item{sigmaE}{double. Variance of environmental values}

\item{Pfall}{double. Treefall probability}

\item{Rgaps}{int. Treefall gaps radius}

\item{Pdeath}{double. Background mortality probability}

\item{Ns}{int. Number of seedlings per cell}

\item{Rdispersal}{int. Dispersal radius in cells}

\item{determinist}{bool. Deterministic or probabilistic vaibility}
}
\value{
A data frame.
}
\description{
simulator
}
\examples{

simulator()

}
