% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulator.R
\name{simulator}
\alias{simulator}
\title{simulator}
\usage{
simulator(grid = 20, Nt = 50, topography = "sinusoidal", Elim = 5,
  amplitude = 0.01, rudgeness = 1, plot = 1, muG = 0, sigmaG = 1,
  muE = 0, sigmaE = 1, Pfall = 0.01, Rgaps = 2, Pdeath = 0.1,
  Ns = 4, Rdispersal = 1, determinist = TRUE)
}
\arguments{
\item{grid}{int.  Number of cells per side of the matrix, should be of size 2^n+1 for the square-diamond algorithm}

\item{Nt}{int. Number of time steps}

\item{topography}{char. Topography generator between 'sinusoidal', 'squarediamond' or 'paracou'}

\item{Elim}{double. Environmental matrix extrme (absolute value)}

\item{amplitude}{double. Amplitude of the sinusoidal functional}

\item{rudgeness}{double. Rugedness parameter}

\item{plot}{int. Plot number between 1 and 15 (a cell size is 3x3m)}

\item{muG}{double. Mean of genetic values}

\item{sigmaG}{double. Variance of genetic values}

\item{muE}{double. Mean of environmental values}

\item{sigmaE}{double. Variance of environmental values}

\item{Pfall}{double. Treefall probability}

\item{Rgaps}{int. Treefall gaps radius}

\item{Pdeath}{double. Background mortality probability}

\item{Ns}{int. Number of seedlings per cell}

\item{Rdispersal}{int. Dispersal radius in cells}

\item{determinist}{bool. Deterministic or probabilistic vaibility}
}
\value{
A data frame.
}
\description{
simulator
}
\examples{

simulator()

}
