% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulator.R
\name{simulator}
\alias{simulator}
\title{Simulator}
\usage{
simulator(
  grid = 20,
  Nt = 50,
  timestep = 30,
  topography = generateTopography(),
  NCI = generateNCI(),
  sigmaGtopo = 1,
  sigmaZtopo = 1,
  sigmaGnci = 26,
  sigmaZnci = 26,
  Pdeath = 0.01325548,
  Ns = 4,
  Rpollination = 1,
  Rdispersion = 1,
  determinist = TRUE
)
}
\arguments{
\item{grid}{int.  Number of cells per side of the matrix.}

\item{Nt}{int. Number of time steps.}

\item{timestep}{int. Time-step length in years.}

\item{topography}{matrix. Topography matrix generated with generateTopography.}

\item{NCI}{matrix. Topography matrix generated with generateNCIsim.}

\item{sigmaGtopo}{double. Variance of genetic values with topography.}

\item{sigmaZtopo}{double. Plasticity of phenotypes with topography.}

\item{sigmaGnci}{double. Variance of genetic values with NCI.}

\item{sigmaZnci}{double. Plasticity of phenotypes with NCI.}

\item{Pdeath}{double. Background mortality probability.}

\item{Ns}{int. Number of seedlings per cell.}

\item{Rpollination}{int. Pollination radius in cells (father to mother).}

\item{Rdispersion}{int. Dispersal radius in cells (mother to seedling).}

\item{determinist}{bool. Deterministic or probabilistic viability.}
}
\value{
A data frame.
}
\description{
Simulator
}
\examples{

simulator()

}
