% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulator.R
\name{simulator}
\alias{simulator}
\title{simulator}
\usage{
simulator(grid = 20, Ngen = 50, muG = 0, sigmaG = 1, muE = 0,
  sigmaE = 1, Elim = 5, seedlings = 4, dispersal = 1,
  gapradius = 2, fallprobability = 0.01, death = 0.1,
  viability_deterministic = TRUE)
}
\arguments{
\item{grid}{int.  Number of cells per side of the environmental matrix}

\item{Ngen}{int. Number of generations}

\item{muG}{double. Mean of genetic values}

\item{sigmaG}{double. Variance of genetic values}

\item{muE}{double. Mean of environmental values}

\item{sigmaE}{double. Variance of environmental values}

\item{Elim}{double. Environmental gradient size}

\item{seedlings}{int. Number of seedlings per cell}

\item{dispersal}{int. Dispersal distance in cells}

\item{gapradius}{int. Gaps radius}

\item{fallprobability}{double. Gaps probability}

\item{death}{double. Death probabilty}

\item{viability_deterministic}{bool. Deterministic or probabilistic vaibility}
}
\value{
A data frame.
}
\description{
simulator
}
\examples{

simulator()

}
