% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{simulatorCpp}
\alias{simulatorCpp}
\title{simulator C++}
\usage{
simulatorCpp(grid = 20L, Ngen = 50L, muG = 0, sigmaG = 1,
  muE = 0, sigmaE = 1, Elim = 5, seedlings = 4L, dispersal = 1L,
  gapradius = 2L, fallprobability = 0.01, death = 0.1,
  viability_deterministic = TRUE)
}
\arguments{
\item{grid}{int. Number of cell per side of the environmental matrix}

\item{Ngen}{int. Number of generations}

\item{muG}{double. Mean of genetic values}

\item{sigmaG}{double. Variance of genetic values}

\item{muE}{double. Mean of environmental values}

\item{sigmaE}{double. Variance of environmental values}

\item{Elim}{double. Environmental gradient size}

\item{seedlings}{int. Number of seedlings per cell}

\item{dispersal}{int. Dispersal distance in cells}

\item{gapradius}{int. Gaps radius}

\item{fallprobability}{double. Gaps probability}

\item{death}{double. Death probabilty}

\item{viability_deterministic}{bool. Deterministic or probabilistic vaibility}
}
\value{
A lsit.
}
\description{
Eco-evolutionary simulator in C++.
}
\examples{
simulatorCpp()

}
