% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{simulatorCpp}
\alias{simulatorCpp}
\title{simulator C++}
\usage{
simulatorCpp(Topo, NCI, grid = 10L, Nt = 50L, sigmaGtopo = 1,
  sigmaZtopo = 1, sigmaGnci = 2.651, sigmaZnci = 2.651,
  Pdeath = 0.01325548, Ns = 4L, Rdispersal = 1L,
  determinist = TRUE)
}
\arguments{
\item{Topo}{matrix. Topography matrix generated with `sinusoidalTopogrpahy`, 
`squareDiamondTopography`, `paracouTopography`.}

\item{NCI}{matrix. NCI matrix generated with `generateNCI`.}

\item{grid}{int. Number of cell per side of the matrix.}

\item{Nt}{int. Number of time steps.}

\item{sigmaGtopo}{double. Variance of genetic values with topography.}

\item{sigmaZtopo}{double. Plasticity of phenotypes with topography.}

\item{sigmaGnci}{double. Variance of genetic values with NCI.}

\item{sigmaZnci}{double. Plasticity of phenotypes with NCI.}

\item{Pdeath}{double. Background mortality probability.}

\item{Ns}{int. Number of seedlings per cell.}

\item{Rdispersal}{int. Dispersal radius in cells.}

\item{determinist}{bool. Deterministic or probabilistic vaibility.}
}
\value{
List of 6 Topography, Atopo, Ztopo, NCI, Anci, Znci.
}
\description{
Eco-evolutionary simulator in C++.
}
\examples{
simulatorCpp(Topo = sinusoidalTopography(grid = 10, Elim = 5, amplitude = 0.01), 
             NCI = generateNCIsim(grid = 10, Nt = 50))

}
