% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{simulatorCpp}
\alias{simulatorCpp}
\title{simulator C++}
\usage{
simulatorCpp(Topography, grid = 10L, Nt = 50L, muG = 0, sigmaG = 1,
  muE = 0, sigmaE = 1, Pfall = 0.01, Rgaps = 2L, Pdeath = 0.1,
  Ns = 4L, Rdispersal = 1L, determinist = TRUE)
}
\arguments{
\item{Topography}{matrix. Environmental matrix}

\item{grid}{int. Number of cell per side of the matrix}

\item{Nt}{int. Number of time steps}

\item{muG}{double. Mean of genetic values}

\item{sigmaG}{double. Variance of genetic values}

\item{muE}{double. Mean of environmental values}

\item{sigmaE}{double. Variance of environmental values}

\item{Pfall}{double. Treefall probability}

\item{Rgaps}{int. Treefall gaps radius}

\item{Pdeath}{double. Background mortality probability}

\item{Ns}{int. Number of seedlings per cell}

\item{Rdispersal}{int. Dispersal radius in cells}

\item{determinist}{bool. Deterministic or probabilistic vaibility}
}
\value{
A lsit.
}
\description{
Eco-evolutionary simulator in C++.
}
\examples{
simulatorCpp(Topography = sinusoidalTopography(grid = 10, Elim = 5, amplitude = 0.01))

}
