% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/squareDiamondTopography.R
\name{squareDiamondTopography}
\alias{squareDiamondTopography}
\title{Topography generator with the square-diamond alogirthm}
\usage{
squareDiamondTopography(grid = 2^6 + 1, rudgeness = 1)
}
\arguments{
\item{grid}{int. Matrix grid size should be of size 2^n+1}

\item{rudgeness}{double. Rugedness parameter}
}
\value{
Environmental matrix of size grid x grid for the simulator
}
\description{
Topography generator with the square-diamond alogirthm
}
\examples{
squareDiamondTopography(5,1)

}
